Generate record constructors with and without an Arena parameter.
The constructors without the Arena parameter will use Arena::ofAuto.

The Arena parameter has been moved to the end of the parameter list (in constructors and field setters).

This is a breaking change (in the field settings). The old `allocate` methods are marked as deprecated.

To prevent name clashes with constructors that already exist in GObject-Introspection data, they are not added when a constructor named `new` already exists.

Fixes #86