Change type of GObject constant
`TYPE_FLAG_RESERVED_ID_BIT` is defined as a gtype with a constant value. I could try to improve the code generator so it will output `new Type(...)` but this seems to be the only case. Moreover, this constant will probably never be used from Java. So I patched it to be a literal number instead of a gtype (which should technically be the same).