Changed toggle ref handling in InstanceCache
Adding and removing GObject toggle refs was done with the generated bindings for these methods in GObject.java, but that will allocate the `toggle-notify` upcall in an arena that is automatically closed when the GObject is garbage-collected. This has now been changed such that the upcall is allocated once (as a static final field) and is then used for all toggle refs.

Improved formatting as well.