Generate static MethodHandle fields
When MethodHandles are generated as static final fields, the JVM will be better able to inline the calls.

The static fields are generated in a private static inner class `MethodHandles`. For interfaces, this is a separate, package-private class, named `[Interface name]MethodHandles`.

Lookup of virtual method calls is done at runtime, so this is not changed. These MethodHandles are cached after the first lookup.