Change bitfield (flag) parameters to EnumSet
This replaces the custom Bitfield base class of all flag types, with Set parameters.

To avoid having to type `Set.of(flag)` when you only want to pass a single flag, overloads are generated for all functions that take flag parameters, so you can pass either a Set or a single flag.