Fixed bugs in the MemoryCleaner class
Removed the `reference` counter field because there can always be only one single instance that has ownership.

Clarified the javadoc comments and improved some formatting.

Moved the call to `g_free()` down, below the calls to `g_boxed_free()` and the specialized free-function, to make it more clear that it is a fallback case.