Implement malloc-based SegmentAllocator
When a memory segment is passed to a native function with full ownership transfer, the native function will free the memory. Until now, java-gi used Arena.global() to allocate this memory, to prevent a double-free, but this is not guaranteed to work. It is better to actually use malloc to allocate the memory.

The malloc function we use is g_try_malloc0. This means the newly allocated memory will be zero-initialized. If the allocation fails, it returns null.

The allocator contains a workaround to be able to allocate zero-length memory segments (common in Java, but not supported by g_malloc).