Improve handling of GValues with boxed types and string arrays
Java-GI will now register boxed types in the TypeCache, so when a GValue contains a boxed value, Java-GI is able to construct the expected Java  proxy class instance.

Also added handling of GStrv GValues, so they are automatically converted from and to Java String[].

The GValue conversion functions will now throw an UnsupportedOperationException when a type cannot be convered.