Fix bug in enum/flags TypeCache
Putting a `null` type into a map that doesn't allow `null` values (when registering an enum type that doesn't exist) caused the class initializers to fail and the JVM to quit with a confusing stacktrace.

This caused CI to fail, because the GH action runs on an older Ubuntu that doesn't have the latest GTK with all types that java-gi tries to cache.