Move toString() method from GObject to GLib Type class
java-gi offers a Type.toString() method that calls g_type_name(), but
because that function is in GObject, it wasn't available for the alias
that java-gi has in GLib. Now that the packages are combined in one
module, it's possible to call a GObject function from the GLib side.

This change uncovered a bug in the metadata functionality, that caused
an immutable attributes map in GirElement. The metadata matcher
requires a mutable map, so this is now fixed.