Prepare to make ProxyInstance.address field private again
The address of a java-gi proxy class must be accessed with the `handle()` method (as specified in the Proxy interface). The actual address is stored in the `address` field of the ProxyInstance class. That field had private visibility until java-gi 0.12, when it was changed to public visibility. I want to revert that, so it is now marked as deprecated, and will be changed to private visibility in a future release.