Change Flags.of() factory method to return a Set
For a Flags (i.e. bitfield) type, the Flags.of() static factory
method assumed the input value represented exactly one flag. This
is now fixed, and will return a Set<Flag> instead.

This is a breaking change for existing users of this method. But
because the old behavior was broken, I expect this will not be a
huge problem.