Update to GNOME 49
Updated the gir-files repository to the latest version, which includes
GNOME 49 gir files for Linux, Windows and macOS.

Because the DesktopAppInfo has moved into a separate library, the
StrvArrayTest testcase didn't work anymore. I removed the testcase,
because there are multiple testcases in gobject-introspection-tests
that test GStrv marshaling (and much more thoroughly).

Likewise, Win32NetworkMonitor is not in Gio anymore, so the metadata
override to skip it in the Java-GI bindings, has been removed.